Camera working

+++ Test_folder/Tho_module/Camera_PC/Transfer_to_PC.py
@@ -1,38 +1,45 @@
 import io
+import socket
+import struct
 import time
 import picamera
-from PIL import Image
-import cv2
-import numpy
 
-def outputs():
+# Connect a client socket to my_server:8000 (change my_server to the
+# hostname of your server)
+client_socket = socket.socket()
+client_socket.connect(('my_server', 8000))
+
+# Make a file-like object out of the connection
+connection = client_socket.makefile('wb')
+try:
+    camera = picamera.PiCamera()
+    camera.resolution = (640, 480)
+    # Start a preview and let the camera warm up for 2 seconds
+    camera.start_preview()
+    time.sleep(2)
+
+    # Note the start time and construct a stream to hold image data
+    # temporarily (we could write it directly to connection but in this
+    # case we want to find out the size of each capture first to keep
+    # our protocol simple)
+    start = time.time()
     stream = io.BytesIO()
-    for i in range(40):
-        # This returns the stream for the camera to capture to
-        yield stream
-        # Once the capture is complete, the loop continues here
-        # (read up on generator functions in Python to understand
-        # the yield statement). Here you could do some processing
-        # on the image...
+    for foo in camera.capture_continuous(stream, 'jpeg'):
+        # Write the length of the capture to the stream and flush to
+        # ensure it actually gets sent
+        connection.write(struct.pack('<L', stream.tell()))
+        connection.flush()
+        # Rewind the stream and send the image data over the wire
         stream.seek(0)
-        pil_image = Image.open(stream)
-        #cv2.imread(img)
-        #img_bytes = stream.read()
-        # Finally, reset the stream for the next capture
-        #print("{}".format(img))
-        open_cv_image = numpy.array(pil_image)
-        open_cv_image = open_cv_image[:, :, ::-1].copy()
-        cv2.imwrite('test.jpg', open_cv_image)
+        connection.write(stream.read())
+        # If we've been capturing for more than 30 seconds, quit
+        if time.time() - start > 30:
+            break
+        # Reset the stream for the next capture
         stream.seek(0)
         stream.truncate()
-
-
-with picamera.PiCamera() as camera:
-    camera.resolution = (640, 480)
-    camera.framerate = 80
-    time.sleep(2)
-    start = time.time()
-    camera.capture_sequence(outputs(), 'jpeg', use_video_port=True)
-    finish = time.time()
-    print('Captured 40 images at %.2ffps' % (40 / (finish - start)))
-
+    # Write a length of zero to the stream to signal we're done
+    connection.write(struct.pack('<L', 0))
+finally:
+    connection.close()
+    client_socket.close()
